Sprint 13. Refactoring FilmDbStorage
